Master

+++ src/Illuminate/Filesystem/Filesystem.php
@@ -1,9 +1,14 @@
 <?php namespace Illuminate\Filesystem;
 
+use ErrorException;
 use FilesystemIterator;
 use Symfony\Component\Finder\Finder;
+use Illuminate\Support\Traits\Macroable;
+use Illuminate\Contracts\Filesystem\FileNotFoundException;
 
 class Filesystem {
+
+	use Macroable;
 
 	/**
 	 * Determine if a file exists.
@@ -22,7 +27,7 @@ class Filesystem {
 	 * @param  string  $path
 	 * @return string
 	 *
-	 * @throws FileNotFoundException
+	 * @throws \Illuminate\Contracts\Filesystem\FileNotFoundException
 	 */
 	public function get($path)
 	{
@@ -37,7 +42,7 @@ class Filesystem {
 	 * @param  string  $path
 	 * @return mixed
 	 *
-	 * @throws FileNotFoundException
+	 * @throws \Illuminate\Contracts\Filesystem\FileNotFoundException
 	 */
 	public function getRequire($path)
 	{
@@ -62,11 +67,12 @@ class Filesystem {
 	 *
 	 * @param  string  $path
 	 * @param  string  $contents
+	 * @param  bool  $lock
 	 * @return int
 	 */
-	public function put($path, $contents)
+	public function put($path, $contents, $lock = false)
 	{
-		return file_put_contents($path, $contents);
+		return file_put_contents($path, $contents, $lock ? LOCK_EX : 0);
 	}
 
 	/**
@@ -110,7 +116,20 @@ class Filesystem {
 
 		$success = true;
 
-		foreach ($paths as $path) { if ( ! @unlink($path)) $success = false; }
+		foreach ($paths as $path)
+		{
+			try
+			{
+				if ( ! @unlink($path))
+				{
+					$success = false;
+				}
+			}
+			catch (ErrorException $e)
+			{
+				$success = false;
+			}
+		}
 
 		return $success;
 	}
@@ -140,6 +159,17 @@ class Filesystem {
 	}
 
 	/**
+	 * Extract the file name from a file path.
+	 *
+	 * @param  string  $path
+	 * @return string
+	 */
+	public function name($path)
+	{
+		return pathinfo($path, PATHINFO_FILENAME);
+	}
+
+	/**
 	 * Extract the file extension from a file path.
 	 *
 	 * @param  string  $path
@@ -159,6 +189,17 @@ class Filesystem {
 	public function type($path)
 	{
 		return filetype($path);
+	}
+
+	/**
+	 * Get the mime-type of a given file.
+	 *
+	 * @param  string  $path
+	 * @return string|false
+	 */
+	public function mimeType($path)
+	{
+		return finfo_file(finfo_open(FILEINFO_MIME_TYPE), $path);
 	}
 
 	/**
@@ -367,7 +408,7 @@ class Filesystem {
 			// If the item is a directory, we can just recurse into the function and
 			// delete that sub-directory otherwise we'll just delete the file and
 			// keep iterating through each file until the directory is cleaned.
-			if ($item->isDir())
+			if ($item->isDir() && ! $item->isLink())
 			{
 				$this->deleteDirectory($item->getPathname());
 			}
